slash-commands-and-actions/requirements.txt
Flask==1.1.1
requests==2.22.0
google-cloud-firestore==1.4.0
pytz==2019.2